Possible #7 workaround
It's unclear what the implications of this commit may be. The MessageHandler#broadcastMessage method exited to prevent false join/leave/swap messages from being broadcasted due to an underlying error such as a server being offline. However, leave messages continue to not be broadcasted resulting in duplicate join messages when the player rejoins (#7). This inconsistency is confusing and leads to the messages not being trusted at all.